Fix warning
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@12729 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/extlib/fbudf/fbudf.cpp
@@ -489,7 +489,7 @@ namespace internal
 			--v->timestamp_date;
 			v->timestamp_time += tenthmsec_in_day + secs;
 		}
-		else if ((v->timestamp_time += secs) >= tenthmsec_in_day)
+		else if ((v->timestamp_time += secs) >= static_cast<unsigned long>(tenthmsec_in_day))
 		{
 			++v->timestamp_date;
 			v->timestamp_time -= tenthmsec_in_day;
